moved fundraiser CRUD bodies to common
