+++ backend/src/api/fundraiser/fundraiser.handlers.ts
@@ -1,14 +1,16 @@
 import { Request, Response } from "express-serve-static-core";
 import {
-  CreateFundraiserBody,
-  FundraiserRouteParams,
-  UpdateFundraiserBody,
-  CreateFundraiserItemBody,
-  UpdateFundraiserItemBody,
-  FundraiserItemRouteParams,
-  CreateAnnouncementBody,
-  DeleteAnnouncementRouteParams,
-} from "./fundraiser.types";
+	FundraiserRouteParams,
+	FundraiserItemRouteParams,
+	DeleteAnnouncementRouteParams,
+} from "./fundraiser/types.ts";
+import {
+	CreateFundraiserBody,
+	UpdateFundraiserBody,
+	CreateFundraiserItemBody,
+	UpdateFundraiserItemBody,
+	CreateAnnouncementBody,
+} from "common";
 import {
   createFundraiser,
   getAllFundraisers,
@@ -133,7 +135,7 @@ export const getFundraiserOrdersHandler = async (
 };
 
 export const createFundraiserHandler = async (
-  req: Request<{}, any, CreateFundraiserBody, {}>,
+  req: Request<{}, any, z.infer<typeof CreateFundraiserBody>, {}>,
   res: Response
 ) => {
   const organization = await getOrganization(req.body.organizationId);
@@ -169,7 +171,7 @@ export const createFundraiserHandler = async (
 };
 
 export const updateFundraiserHandler = async (
-  req: Request<FundraiserRouteParams, any, UpdateFundraiserBody, {}>,
+  req: Request<FundraiserRouteParams, any, z.infer<typeof UpdateFundraiserBody>, {}>,
   res: Response
 ) => {
   const fundraiser = await getFundraiser(req.params.id);
@@ -212,7 +214,7 @@ export const updateFundraiserHandler = async (
 };
 
 export const createFundraiserItemHandler = async (
-  req: Request<FundraiserRouteParams, any, CreateFundraiserItemBody, {}>,
+  req: Request<FundraiserRouteParams, any, z.infer<typeof CreateFundraiserItemBody>, {}>,
   res: Response
 ) => {
   const fundraiser = await getFundraiser(req.params.id);
@@ -255,7 +257,7 @@ export const createFundraiserItemHandler = async (
 };
 
 export const updateFundraiserItemHandler = async (
-  req: Request<FundraiserItemRouteParams, any, UpdateFundraiserItemBody, {}>,
+  req: Request<FundraiserItemRouteParams, any, z.infer<typeof UpdateFundraiserItemBody>, {}>,
   res: Response
 ) => {
   const fundraiser = await getFundraiser(req.params.fundraiserId);
@@ -298,7 +300,7 @@ export const updateFundraiserItemHandler = async (
 };
 
 export const createAnnouncementHandler = async (
-  req: Request<FundraiserRouteParams, any, CreateAnnouncementBody, {}>,
+  req: Request<FundraiserRouteParams, any, z.infer<typeof CreateAnnouncementBody>, {}>,
   res: Response
 ) => {
   const fundraiser = await getFundraiser(req.params.id);
